+++ CHANGELOG.md
@@ -1,15 +1,19 @@
 # Changelog for `toml-env`
 
+## v0.3.0
+
+- Improved api for `Args::map_env` using `IntoIterator`.
+
 ## v0.2.0
 
-+ Support for mapping environment variables into the config (`X => y.z`).
+- Support for mapping environment variables into the config (`X => y.z`).
 
 ## v0.1.1
 
 ### Documentation
 
-+ Fix example in README.
+- Fix example in README.
 
 ## v0.1.0
 
-+ Initial release.
+- Initial release.